foundation for dynamo test approach, update support files, add test helpers
